added unit tests

+++ Simple.Inputs/Inputs/EmploymentInput.cs
@@ -1,3 +1,3 @@
 namespace Simple.Model.Inputs;
 
-public record EmploymentInput(string Company, uint MonthsOfExperience, double Salary, DateTime StartDate, DateTime? EndDate);
+public record EmploymentInput(string Company, short MonthsOfExperience, double Salary, DateTime StartDate, DateTime? EndDate);
